# src/main/java/br/com/ForumHub/infra/springDoc/api-docs.json
{"openapi":"3.1.0","info":{"title":"ForumHub API","description":"API REST de um fórum de discussão - Alura ONE Oracle","contact":{"name":"MP","email":"mp@email.com"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Cadastro","description":"Registro de novos usuários"},{"name":"Tópicos","description":"Gerenciamento de tópicos do fórum"},{"name":"Autenticação","description":"Login e geração de token JWT"}],"paths":{"/topicos/{id}":{"get":{"tags":["Tópicos"],"summary":"Busca tópico por ID","description":"Não requer autenticação","operationId":"buscarPorId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Tópico encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicoDetalheDTO"}}}},"404":{"description":"Tópico não encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicoDetalheDTO"}}}}},"security":[{"Bearer Authentication":[]}]},"put":{"tags":["Tópicos"],"summary":"Atualiza um tópico","description":"Requer autenticação. Apenas o autor pode atualizar.","operationId":"atualizar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TopicoFormDTO"}}},"required":true},"responses":{"200":{"description":"Tópico atualizado com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicoDetalheDTO"}}}},"400":{"description":"Tópico duplicado ou dados inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicoDetalheDTO"}}}},"403":{"description":"Usuário não é o autor do tópico","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicoDetalheDTO"}}}},"404":{"description":"Tópico não encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicoDetalheDTO"}}}},"401":{"description":"Não autenticado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicoDetalheDTO"}}}}},"security":[{"Bearer Authentication":[]}]},"delete":{"tags":["Tópicos"],"summary":"Deleta um tópico","description":"Requer autenticação","operationId":"deletar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"Tópico deletado com sucesso"},"404":{"description":"Tópico não encontrado"},"401":{"description":"Não autenticado"}},"security":[{"Bearer Authentication":[]}]}},"/topicos":{"get":{"tags":["Tópicos"],"summary":"Lista todos os tópicos","description":"Não requer autenticação","operationId":"listar","responses":{"200":{"description":"Lista de tópicos retornada com sucesso","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TopicoDetalheDTO"}}}}}},"security":[{"Bearer Authentication":[]}]},"post":{"tags":["Tópicos"],"summary":"Cadastra um novo tópico","description":"Requer autenticação. Título e mensagem devem ser únicos.","operationId":"cadastrar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TopicoFormDTO"}}},"required":true},"responses":{"201":{"description":"Tópico criado com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicoDetalheDTO"}}}},"400":{"description":"Tópico duplicado ou dados inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicoDetalheDTO"}}}},"401":{"description":"Não autenticado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopicoDetalheDTO"}}}}},"security":[{"Bearer Authentication":[]}]}},"/topicos/{id}/respostas":{"get":{"tags":["Tópicos"],"summary":"Lista respostas de um tópico","description":"Não requer autenticação","operationId":"listarRespostas","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Lista de respostas retornada com sucesso","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RespostaDetalheDTO"}}}}},"404":{"description":"Tópico não encontrado","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RespostaDetalheDTO"}}}}}},"security":[{"Bearer Authentication":[]}]},"post":{"tags":["Tópicos"],"summary":"Cadastra uma resposta em um tópico","description":"Requer autenticação","operationId":"cadastrarResposta","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RespostaFormDTO"}}},"required":true},"responses":{"201":{"description":"Resposta criada com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RespostaDetalheDTO"}}}},"404":{"description":"Tópico não encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RespostaDetalheDTO"}}}},"400":{"description":"Dados inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RespostaDetalheDTO"}}}},"401":{"description":"Não autenticado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RespostaDetalheDTO"}}}}},"security":[{"Bearer Authentication":[]}]}},"/register":{"post":{"tags":["Cadastro"],"summary":"Cadastra um novo usuário","description":"Requer login (e-mail) e senha. O login deve ser único.","operationId":"cadastrar_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DadosCadastroUsuario"}}},"required":true},"responses":{"201":{"description":"Usuário cadastrado com sucesso","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Login já existe ou dados inválidos","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/login":{"post":{"tags":["Autenticação"],"summary":"Efetua login","description":"Retorna um token JWT para autenticação nas demais requisições","operationId":"efetuarLogin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DadosAutenticacao"}}},"required":true},"responses":{"200":{"description":"Login realizado com sucesso, retorna token JWT","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DadosTokenJWT"}}}},"401":{"description":"Credenciais inválidas","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DadosTokenJWT"}}}}}}}},"components":{"schemas":{"TopicoFormDTO":{"type":"object","description":"Dados necessários para criar ou atualizar um tópico","properties":{"titulo":{"type":"string","description":"Título do tópico (deve ser único)","example":"Como usar Optional no Java?","minLength":1},"mensagem":{"type":"string","description":"Corpo/mensagem do tópico (deve ser único)","example":"Gostaria de entender como funciona o Optional...","minLength":1},"curso":{"type":"string","description":"Categoria/curso do tópico","example":"Java","minLength":1}},"required":["curso","mensagem","titulo"]},"TopicoDetalheDTO":{"type":"object","description":"Detalhes de um tópico retornado pela API","properties":{"id":{"type":"integer","format":"int64","description":"ID do tópico","example":1},"titulo":{"type":"string","description":"Título do tópico","example":"Como usar Optional no Java?"},"mensagem":{"type":"string","description":"Corpo/mensagem do tópico","example":"Gostaria de entender como funciona o Optional..."},"dataCriacao":{"type":"string","format":"date-time","description":"Data de criação do tópico","example":"2025-07-14T22:15:31"},"status":{"type":"string","description":"Status do tópico","example":"ATIVO"},"autor":{"type":"string","description":"E-mail do autor do tópico","example":"usuario@email.com"},"curso":{"type":"string","description":"Categoria/curso do tópico","example":"Java"}}},"RespostaFormDTO":{"type":"object","description":"Dados para cadastrar uma resposta em um tópico","properties":{"mensagem":{"type":"string","description":"Conteúdo da resposta","example":"Use Optional.ofNullable() para tratar valores nulos","minLength":1}},"required":["mensagem"]},"RespostaDetalheDTO":{"type":"object","description":"Detalhes de uma resposta retornado pela API","properties":{"id":{"type":"integer","format":"int64","description":"ID da resposta","example":1},"mensagem":{"type":"string","description":"Conteúdo da resposta","example":"Use Optional.ofNullable()..."},"dataCriacao":{"type":"string","format":"date-time","description":"Data de criação da resposta","example":"2025-07-15T10:30:00"},"autor":{"type":"string","description":"E-mail do autor da resposta","example":"usuario@email.com"},"solucao":{"type":"boolean","description":"Se a resposta é a solução aceita","example":false}}},"DadosCadastroUsuario":{"type":"object","description":"Dados necessários para cadastro de novo usuário","properties":{"login":{"type":"string","description":"E-mail do usuário (usado como login)","example":"usuario@email.com","minLength":1},"senha":{"type":"string","description":"Senha do usuário","example":"123456","minLength":1}},"required":["login","senha"]},"DadosAutenticacao":{"type":"object","description":"Dados necessários para autenticação","properties":{"login":{"type":"string","description":"E-mail do usuário","example":"usuario@email.com"},"senha":{"type":"string","description":"Senha do usuário","example":"123456"}}},"DadosTokenJWT":{"type":"object","description":"Resposta com o token JWT gerado","properties":{"token":{"type":"string","description":"Token JWT para autenticação nas requisições","example":"eyJhbGciOiJIUzI1NiJ9..."}}}},"securitySchemes":{"Bearer Authentication":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}